src/Casa/requirements.txt
anyio==4.10.0
asgiref==3.10.0
babel==2.17.0
backrefs==5.9
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
csscompressor==0.9.5
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
ghp-import==2.1.0
h11==0.16.0
htmlmin2==0.1.13
httpcore==1.0.9
httpx==0.27.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsmin==3.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.19
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
mkdocs_puml==2.3.0
msgpack==1.1.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.4.0
Pygments==2.19.2
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.5
rich==13.9.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
